==> Notebooks/1. Cargar Archivos Luxury_Beauty.ipynb <==
reviewerID - ID of the reviewer, e.g. A2SUAM1J3GNN3B

asin - ID of the product, e.g. 0000013714

reviewerName - name of the reviewer

vote - helpful votes of the review

style - a disctionary of the product metadata, e.g., "Format" is "Hardcover"

reviewText - text of the review

overall - rating of the product

summary - summary of the review

unixReviewTime - time of the review (unix time)

reviewTime - time of the review (raw)

image - images that users post after they have received the product

import pandas as pd
import io
import json
import gzip
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
# Configurar pandas para mostrar todas las columnas
pd.set_option('display.max_columns', None)


pip install pandas openpyxl

### Carga la data de reviews

data = []
with gzip.open('Luxury_Beauty.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))

# Longitud total de la lista, este numero equivale al numero total de productos
print(len(data))

# Primera fila de la lista
print(data[0])

# Convierte la lista en un dataframe de pandas

df1 = pd.DataFrame.from_dict(data)

print(len(df1))

df1.head(5)

### Carga la meta data de reviews

data = []
with gzip.open('meta_Luxury_Beauty.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))

# Longitud total de la lista, este numero equivale al numero total de productos
print(len(data))

# Primera fila de la lista
print(data[0])

# Convierte la lista en un dataframe de pandas

df2 = pd.DataFrame.from_dict(data)

print(len(df2))

# Guardar el DataFrame en formato Excel
#df2.to_excel('nombre_del_archivo_salida.xlsx', index=False, engine='openpyxl')

print("Archivo convertido y guardado con éxito.")

result = pd.merge(df1,df2, on = "asin")

result.head(2)

result.info()

data = result

# Lista de columnas a eliminar
columns_to_drop = ['category','main_cat','feature','vote','style', 'reviewerName', 'image','title','date','imageURL','imageURLHighRes','also_view','also_buy','fit','details','similar_item','tech1','tech2','description']

# Eliminar columnas
data = data.drop(columns=columns_to_drop, errors='ignore')

data.dropna(inplace=True)

# Extraer todos los números y comas del principio de la cadena
extracted_series = data['rank'].str.extract(r'([\d,]+)')[0]

# Convertir "nan" (cadena) a np.nan
extracted_series.replace("nan", np.nan, inplace=True)


# Eliminar comas y convertir a int
data['extracted_rank'] = extracted_series.str.replace(',', '').fillna('0').astype(int)


# Rellenar valores NaN con 0, eliminar comas y convertir a entero
data['extracted_rank'] = extracted_series.fillna(0).astype(str).str.replace(',', '').astype(int)

# Extraer el primer valor cuando hay un rango de precios
data['price'] = data['price'].str.extract(r'(\d+.\d+)')[0]

# Eliminar el símbolo $ y las comas, luego convertir a float
data['price'] = data['price'].str.replace('$', '', regex=True).str.replace(',', '').astype(float)

# Convertir la columna unixReviewTime a un formato de fecha
data['reviewDate'] = pd.to_datetime(data['unixReviewTime'], unit='s').dt.date

data = data[data['verified'] == True]

data.info()

data.head(2)

# Lista de columnas a eliminar
columns_to_drop = ['reviewTime','unixReviewTime','rank','verified']

# Eliminar columnas
data = data.drop(columns=columns_to_drop, errors='ignore')

# Convertir reviewDate a tipo de dato datetime
data['reviewDate'] = pd.to_datetime(data['reviewDate'])

data.head(2)

data.info()

# 1. Entendimiento Básico
print(data.shape)
#print(data.info())
#print(data.head())
#print(data.tail())

unique_values_freq = data['brand'].value_counts()
print(unique_values_freq)

# 2. Estadísticas Descriptivas
print(data.describe())
print(data.describe(include=[np.object]))

# 3. Manejo de Valores Faltantes
missing_values = data.isnull().sum()
print(missing_values)

data.dropna(inplace=True)

# 3. Manejo de Valores Faltantes
missing_values = data.isnull().sum()
print(missing_values)

print(data.shape)

# Guardar el DataFrame en un archivo CSV
data.to_csv('Luxury_Beauty.csv', index=False)  # 'index=False' para no guardar el índice del DataFrame

import re

# Función para eliminar/reemplazar caracteres no permitidos
def clean_illegal_chars(val):
    if isinstance(val, str):  # Solo procesar cadenas
        # Eliminar caracteres no permitidos
        
        val = re.sub(r'[^\x09\x0A\x0D\x20-\uD7FF\uE000-\uFFFD\u10000-\u10FFFF]', '', val)
    return val

# Aplicar la función a todo el DataFrame
df_cleaned = data.applymap(clean_illegal_chars)

# Guardar el DataFrame en un archivo Excel
df_cleaned.to_excel('Luxury_Beauty.xlsx', index=False, engine='openpyxl')
