==> Notebooks/4. Unificación DFs.ipynb <==
import pandas as pd
import json
import gzip

# Leer el archivo CSV
df1 = pd.read_csv('AMAZON_FASHION.csv')

df1['category'] = 'FASHION'

df1.shape

# Leer el archivo CSV
df2 = pd.read_csv('Luxury_Beauty.csv')

df2['category'] = 'LUXURY_BEAUTY'

df2.shape

df2.head(3)

### Carga la data de reviews

data = []
with gzip.open('Final_Clothing_Shoes_and_Jewelry.json.gz') as f:
    for l in f:
        data.append(json.loads(l.strip()))

# Longitud total de la lista, este numero equivale al numero total de productos
print(len(data))

# Primera fila de la lista
print(data[0])

# Convierte la lista en un dataframe de pandas

df3 = pd.DataFrame.from_dict(data)

list_of_dicts = df3.iloc[0].tolist()

# Convertir la lista de diccionarios en un DataFrame
df = pd.DataFrame(list_of_dicts)

df['category'] = 'CLOTHING_SHOES'

#Convertir el formato de la fecha
df['reviewDate'] = pd.to_datetime(df['reviewDate'], unit='ms')
df['reviewDate'] = df['reviewDate'].dt.strftime('%Y-%m-%d')

df.head(3)

# Unir verticalmente los DataFrames
df_unido = pd.concat([df1, df2,df], axis=0, ignore_index=True)

df_unido.shape

# Guardar el DataFrame en un archivo CSV
df_unido.to_csv('DataFinal_Amazon.csv', index=False)  # 'index=False' para no guardar el índice del DataFrame

# Convertir el DataFrame a una cadena JSON
json_string = df_unido.to_json(orient='records')

# Comprimir y guardar el JSON como .json.gz
with gzip.open('DataFinal_Amazon.json.gz', 'wt') as f:
    f.write(json_string)
